Committed interfaces to replace implementations of two classes (BaseAudio, GameFontManager).  The new interfaces are of the same names as the implementations, and the implementations changed to the name "Defaultxxx" where xxx is the name of the interface.  Also encapsulated the state of BaseGame.  This refactoring will continue until BaseGame relies solely on interfaces, not implementations.  Then, Game and GameObject will be refactored to make them both rely on encapsulation and interfaces, not implementations (note that some behavior is implementation specific: this will be kept (such as the default setting up of implementations of engines such as bsInput, etc)).  Finally, both Game and GameObject will become interfaces.
GameObject will become a sub-interface of Game that adds the functionality to be added to the GameEngine.  In this fashion, Games will be able to be added to the GameEngine with minimal change (simply implement the GameObject interface replacement, and add functionality for the methods).

Once both become interfaces, GameEngine and GameLoader must be refactored to use only these interfaces, and not reference implementation-specific behavior.