Fixed a slight bug in the DefaultBaseAudio class.  Completely encapsulated the state of the Game class, refactoring the classes that used it.  Got rid of a useless method in the ErrorNotificationDialog class, and refactored an existing method to be message based, not state based.  Made the GTGE version string appear in BaseGame, not Game itself.
The next step is to remove private methods in Game, replacing with either public or protected as needed.  Then, make sure its functionality is sound.  Finally, Game will have its public methods pulled up into an interface.  Once Game is an interface, an interface for the GameObject class (how it differs such that it can interact with the GameEngine interface) will be created.  A combination interface for creating Games that can interact with the GameEngine interface will also be provided for convenience.  In this manner, subclasses inheriting from Game will be able to add the GameEngine compatible interface and behave as Games that also exhibit GameObject behavior, that is, they can interact with the GameEngine.