Complete - task : Refactor, if needed, the GameFontManager class.
Deleted the GameFontManager class - users are better served with making their own maps if they need maps of fonts, and using the proper constructors directly.  This also removes the reliance of GTGE on its own GameFont classes, so that it does not care that its GameFont instances are actually provided and supported classes via the framework instead of user provided classes that are unknown to GTGE at compile time.