Incomplete - task : Refactor, if needed, the SpriteGroup class.
SpriteGroup renamed to BasicSpriteGroup to allow the interface "SpriteGroup" to be created.

SpriteGroup now uses a List of Sprites, not a Sprite array, for implementation simplicity.  Users may set their own list implementation if an ArrayList is not sufficient for their needs (LinkedLists would perform better if no Comparator is needed, for example, because Iterators are used throughout for basic operations).